Remove old files, move ClaimInventory.js
